implement player spawner function

+++ game1/src/ConfigParser.h
@@ -23,7 +23,7 @@ public:
 
 	struct ShapeConfing 
 	{
-		int shapeRadius{};
+		float shapeRadius{};
 		int collisionRadius{};
 		int outlineColor[3]{};
 		int outlineThickness{};
